+++ src/resolvers/Album.js
@@ -1,6 +1,14 @@
 const Scraper = require('../helpers/scraper');
 
 module.exports = {
+    getAlbums: async (args) => {
+        const band = args.band || '';
+        const title = args.title || '';
+        const genre = args.genre || '';
+        const type = args.type || ''
+        const start = args.start || 0;
+        return await Scraper.getDiscs(band, title, genre, type, start);
+    },
     getAlbum: async (args) => await Scraper.getDisc(args.id),
     getAlbumSongs: async (args) => await Scraper.getDiscSongs(args.id),
     getAlbumReviews: async (args) => await Scraper.getDiscReviews(args.id),